+++ CHANGELOG.md
@@ -18,6 +18,9 @@
 ### :rocket: New Feature
 - `preview`
   - [#1356](https://github.com/wix-incubator/rich-content/pull/1356) support video in gallery
+  - [#1369](https://github.com/wix-incubator/rich-content/pull/1369) preview content examples in storybook
+- `storybook`
+  - [#1369](https://github.com/wix-incubator/rich-content/pull/1369) preview content examples
 - `giphy`
   - [#1358](https://github.com/wix-incubator/rich-content/pull/1358) gi preview modal empty state
 - `ricos-viewer`